% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_multiPop}
\alias{x_multiPop}
\title{dPCR sample w/ >=3 populations}
\format{
A numeric vector with 1814 droplet fluorescence
}
\source{
\url{https://github.com/Gromgorgel/ddPCR/blob/master/Dataset.zip}
}
\usage{
x_multiPop
}
\description{
The reaction with ID 373 in the \code{Dataset.zip} provided in the repository from Lievens et. al. (2017)
}
\keyword{datasets}
