% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_onePop}
\alias{x_onePop}
\title{dPCR sample w/ 1 population}
\format{
A numeric vector with 8000 droplet fluorescence
}
\usage{
x_onePop
}
\description{
Simulated dataset with very high presence of rain and true mean copies per partition of 0.1.
}
\keyword{datasets}
