% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_twoPop}
\alias{x_twoPop}
\title{dPCR sample w/ 2 populations}
\format{
A numeric vector with 10254 droplet fluorescence
}
\source{
\url{https://github.com/Gromgorgel/ddPCR/blob/master/Dataset.zip}
}
\usage{
x_twoPop
}
\description{
The reaction with ID 9 in the \code{Dataset.zip} provided in the repository from Lievens et. al. (2017)
}
\keyword{datasets}
