% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateConc.R
\name{calculateConc}
\alias{calculateConc}
\title{Target copies estimation}
\usage{
calculateConc(nneg, ntotal, volSamp, volDrp)
}
\arguments{
\item{nneg}{numeric, negative droplet count}

\item{ntotal}{numeric, total droplet count}

\item{volSamp}{numeric, sample volume in microliter}

\item{volDrp}{numeric, droplet (or partition) volume in nanoliter}
}
\description{
Mean target copies per partition (lambda) is derived using Poisson distribution as lambda = -ln(nneg / ntot). Target copies in sample is then calculated as conc = lambda * volSamp/(volDrp * 1000).
}
