% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPopPCR.R
\name{printSummaryConc}
\alias{printSummaryConc}
\title{Print result summary of popPCR}
\usage{
printSummaryConc(result.popPCR)
}
\arguments{
\item{result.popPCR}{returned value of popPCR()}
}
\description{
Print result summary of popPCR
}
