% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPopPCR.R
\name{printSummaryFit}
\alias{printSummaryFit}
\title{Print fitted mixture model estimates from popPCR}
\usage{
printSummaryFit(result.popPCR)
}
\arguments{
\item{result.popPCR}{returned value of popPCR()}
}
\description{
Print fitted mixture model estimates from popPCR
}
